Force selection of libverto-tevent
The only known libverto user in XCP-ng dom0 is gssproxy, whose code
does not care about which libverto backend will be used, so the choice
of which one to use is delegated at runtime to libverto.

libverto has a "default backend" configured, which will be used it
available, and this is libverto-tevent in this version.  Having this
package installed is the only way to ensure the same lib will be used
by gssproxy (and any other similar software, if any), regardless of
whether libverto-libevent is also installed.

This is still not perfect, as having libverto-glib installed
apparently causes libverto-tevent (for some reason) to also load
libglib.so, but should be good enough as no supported package uses the
latter.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>